Move /translations to /shared/translations folder in WebApp
