+++ application/account-management/WebApp/shared/translations/i18n.config.json
