+++ application/back-office/WebApp/lingui.config.ts
@@ -1,6 +1,6 @@
 import type { LinguiConfig } from "@lingui/conf";
 import { formatter } from "@lingui/format-po";
-import { locales, pseudoLocale, sourceLocale } from "./translations/i18n";
+import { locales, pseudoLocale, sourceLocale } from "./shared/translations/i18n";
 
 const config: LinguiConfig = {
   locales,
@@ -8,7 +8,7 @@ const config: LinguiConfig = {
   pseudoLocale,
   catalogs: [
     {
-      path: "<rootDir>/translations/locale/{locale}",
+      path: "<rootDir>/shared/translations/locale/{locale}",
       include: ["<rootDir>"],
       exclude: ["**/node_modules/**", "**/dist", "**/*.d.ts", "**/*.test.*", "**/.*"]
     }
